Add ManagementModelStub class
